logout disabled, barracks repaired, get buildings repaired

+++ src/main/resources/db/migration/V4__troops.sql
@@ -4,7 +4,7 @@ level INT,
 hp INT,
 attack INT,
 defence INT,
-started_at INT,
-finished_at INT,
+started_at BIGINT,
+finished_at BIGINT,
 FOREIGN KEY(kingdom_id)
 REFERENCES kingdoms(id));
